fix bug #1

+++ scripts/testeAssembly.py
@@ -65,17 +65,18 @@ def clearTestDir(testDir):
 
     if f is not False:
             for l in f:
-                if l.strip()[0]!='#':
-                    par   = l.rstrip().split();
-                    name  = par[0]
-                    nTest = int(par[1])
-                    for i in range (0, nTest):
-                        nameTest   = name + str(i)
-                        ramEndSimu = pwd + "tst/" + name + "/" + nameTest + RAM_END_SIMU_FILE
-                        try:
-                            os.remove(ramEndSimu)
-                        except:
-                            pass
+                if len(l.strip()):
+                    if l.strip()[0]!='#':
+                        par   = l.rstrip().split();
+                        name  = par[0]
+                        nTest = int(par[1])
+                        for i in range (0, nTest):
+                            nameTest   = name + str(i)
+                            ramEndSimu = pwd + "tst/" + name + "/" + nameTest + RAM_END_SIMU_FILE
+                            try:
+                                os.remove(ramEndSimu)
+                            except:
+                                pass
     else:
         return(-1)
 
@@ -143,19 +144,20 @@ def compareFromTestDir(testDir):
 
     if f is not False:
         for l in f:
-            if l.strip()[0]!='#':
-                par   = l.rstrip().split();
-                name  = par[0]
-                nTest = int(par[1])
-                for i in range (0, nTest):
-                        nameTest   = name + str(i)
-                        ramEnd     = pwd + "/tst/" + name + "/" + name + "{}".format(i) + RAM_END_FILE
-                        ramEndSimu = pwd + "/tst/" + name + "/" + nameTest + RAM_END_SIMU_FILE
-                        if(os.path.isfile(ramEnd) and os.path.isfile(ramEndSimu)):
-                            rtn = compareRam(nameTest, ramEnd, ramEndSimu)
-                            r = "True" if rtn else "False"
-                            result = {'name':name, 'resultado':rtn, 'teste':i}
-                            log.append(result)
+            if len(l.strip()):
+                if l.strip()[0]!='#':
+                    par   = l.rstrip().split();
+                    name  = par[0]
+                    nTest = int(par[1])
+                    for i in range (0, nTest):
+                            nameTest   = name + str(i)
+                            ramEnd     = pwd + "/tst/" + name + "/" + name + "{}".format(i) + RAM_END_FILE
+                            ramEndSimu = pwd + "/tst/" + name + "/" + nameTest + RAM_END_SIMU_FILE
+                            if(os.path.isfile(ramEnd) and os.path.isfile(ramEndSimu)):
+                                rtn = compareRam(nameTest, ramEnd, ramEndSimu)
+                                r = "True" if rtn else "False"
+                                result = {'name':name, 'resultado':rtn, 'teste':i}
+                                log.append(result)
     else:
         return(-1, result)
